Make openreviews generate quartiles instead of just median.
By reporting 1,2,3 quartiles instead of just the 2nd (median) we can get a
better insight into what the distribution looks like. For example if
the median is good, but the 1st quartile is very low and the 3rd is very
high it means we have some patches would be waiting much longer then
others, but our average(median) may still be good.

Change-Id: I3427e562ab39dc6915ac0d2e42347e0d6e0123bc