Stop ignoring jenkins -1 votes
One of the age calculations is the time patches have been up without any
-1 or -2 votes.  This calculation basically just doesn't reset the timer
if a reviewer rebases their change, or fixes some things before anyone
points it out.

Originally I ignored jenkins here.  It really wasn't a good idea.  If
jenkins gives a -1 vote, people are not likely to go review it, so those
votes should be included here.

Change-Id: Iec8deac21a2cfc07bf23e0d0d75e8ea0a27646af