Fix disagreement percentage calculation
Previously, a disagreement was defined as a core reviewer giving a
negative vote after the person had given a positive vote.  The
percentage was then the number of disagreements / all positive votes.

Later, disagreements were extended to cover the opposite case: when a
core member gives a positive vote after a person had given a negative
vote.

The percentage calculation was not updated to reflect this new
definition.  It still showed disagreements as a percentage of the
positive votes, instead of all of them (+ and -).

Change-Id: I6392055e5d95dfece86cdd427f44e258566e771a