Remove received reviews from output.
Remove the number of reviews received from the reviewers report output.
I think this information is valuable, but the values we're getting are
not accurate.  The problem is that if a patch gets rebased, the trivial
rebase detection will automatically re-apply reviews.  Those are getting
counted even though they aren't new reviews being received.

Change-Id: I3f1c620c225c1d1d3f2a137b095bac847a353238