Support reviewers stats for stable branches
Adds a "-s" option that lets you specify the name of a stable branch
(for example "-s havana") instead of specifying a project (-p) or all
projects (-a).

reviewers.py will then use a specific stable.json file that defines the
set of projects stable-maint-core has authority on, as well as the
contents of that core team.

Note: -s disables result caching since the search is made on
specific branches, which would pollute the general purpose cache.

Change-Id: Idd03f51c98e006972bc93e3ce11e052cd1cc26aa