Update for Gerrit 2.8
Some of the review types returned in the Gerrit JSON data have
changed, so reviewstats wasn't finding any reviews.  This also
adds support for the new Workflow votes so new approvals will
show up correctly.

Note that the output from reviewstats still seems to be rather
bogus because the review timestamps appear to have all been updated
in the upgrade, but I think reviewstats is processing the data as
intended with this change.

Change-Id: I6cccf427a0ff875fdbc59f8c8d83a37db567f786