Encode to utf-8 before printing
When redirecting the output of openreviews to a file, if there are
utf-8 characters in, for example, a commit message, openreviews
will raise an exception because it tries to encode as ascii by
default.

This change explicitly encodes basestring output to utf-8 so it
doesn't matter what the encoding of the output stream is.

Change-Id: I581c1a98bdbbbfb17ccf33c59295d6fbf8729ab9