Document installation from repository
The reviewstats library cannot be installed from PyPI as it is, as
a tag is needed, or at least a package download URL must be
specified. A user can still install the library and the utilities
by cloning the repo and running the setup script, and therefore
this process is documented rather than the erroneous pip command.

Change-Id: I342fa0f68242316e7ef7bb5959118fdd32d10b9e