Make local cache more permanent.
With no cache, this code has to download the entire review history for
a project, which takes a really long time, not to mention it's
generally wasteful.

There was code to keep a local cache, but it expired after an hour.
It would make quick repeated runs go faster, but for the stats I keep
posted for OpenStack projects this didn't help.  I regenerate
everything less often.  My stats site is basically broken because of
how terribly inefficient this is.

This updated code keeps the cache around without expiring it.
Instead, it just updates the cache until it hits the first change that
hasn't changed since our cached version and stops.  This should make
things work much better and allow my stats site to keep working.

Change-Id: I784b99dfabeeca81c779dcfda15f1f20bb736957