Stop hard coding core team member lists
All of the project json files had lists of the core review team
members.  Since this was originally written, gerrit has added a REST
API and this information is available through that API.  Get the info
from there instead when we can.

There are a few project files I didn't convert because the equivalent
gerrit team wasn't immediately obvious.

Change-Id: Ie7ae19d9a58a5148002676604fd6962ea1438786
Signed-off-by: Russell Bryant <rbryant@redhat.com>