Gerrit 2.11 compatibility
Gerrit 2.11 no longer seems to support the sortKey method of
skipping changes.  Instead is provides a --start option that takes
a number of changes to skip.

See: https://review.openstack.org/Documentation/cmd-query.html

In addition, this required a change to the format of the changes
dict.  It turns out we were overwriting some changes if two had the
same id.  This is common in stable branch cherry picks.

In order to avoid this, I used a tuple of id, project, branch
as the key for the changes dict.  This does break compatibility
with existing cache files, so a check was added to clear old
caches.

Both of these changes were required together because otherwise we
end up with a smaller number of items in changes than we should
have, and the --start argument is incorrect so we start double
processing changes, which breaks the loop because it looks like
the change was already cached.
Change-Id: If5ab46ce3367b3790e0db1d8bb1ed18e39202ad9