Optimize core list retrieval in reviewers
The list of core team members was being retrieved for every review
for each patchset. The utils.get_core_team does some caching to make
sure an API call wasn't made for each of these calls, but since the
project never changes for the review we can save a few cycles by not
making the call if it's not necessary.

This moves the retrieval of the core list before the loop through
reviews to make sure it's only called once per patchset.

Change-Id: Iac9f3b6eb71131eccce58a60656865eeb77b8fda