reviewers: make --project and --stable mutually exclusive
When using the reviewers command to get stats on stable branch
reviews, I always think I can specify --stable and --project nova
to get nova stable reviews, but that's not how the command
works and the --project option is ignored resulting in a report
which I don't expect, i.e. there are non-nova reviewers in the
report.

This makes --project and --stable mutually exclusive to avoid that
confusion.

Change-Id: Ie9938f5e68f27b9a5a58635be2185365f041136f