Various updates to pass CI
This cleans up and updates the project configuration to get it up to
current requirements and able to pass CI. Raises some requirements
versions and job settings.

Most notable is probably switching from the "*-no-constraints" Python
jobs that are not really maintained to the common py3 template. This
will run unit tests with additional py3 versions to cover current
OpenStack required versions, as well as using the global upper
constraints. This way the project will track what is being done with
other OpenStack projects.

Change-Id: I390e2a55f475f33a03d393c53a6f79af7a80eaf3
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>